6418 | 786 files | M DumpSource2/commands.txt,  M DumpSource2/convars.txt,  D DumpSource2/schemas/animationsystem/AnimPoseControl.h,  M DumpSource2/schemas/animationsystem/CMoodVData.h,  M DumpSource2/schemas/animationsystem/MoodAnimation_t.h,  M DumpSource2/schemas/animgraphlib/CAnimInputDamping.h,  M DumpSource2/schemas/animgraphlib/CBlendUpdateNode.h,  M DumpSource2/schemas/animgraphlib/CRagdollAnimTag.h,  M DumpSource2/schemas/animgraphlib/CTargetSelectorUpdateNode.h,  M DumpSource2/schemas/animgraphlib/CTargetWarpUpdateNode.h,  M DumpSource2/schemas/animgraphlib/DampingSpeedFunction.h,  M DumpSource2/schemas/animlib/CNmBlendTaskBase.h,  D DumpSource2/schemas/animlib/CNmBoneMask.h,  D DumpSource2/schemas/animlib/CNmBoneMask__SerializedData_t.h,  D DumpSource2/schemas/animlib/CNmBoneMask__WeightInfo_t.h,  M DumpSource2/schemas/animlib/CNmCachedPoseReadTask.h,  M DumpSource2/schemas/animlib/CNmCachedPoseWriteTask.h,  M DumpSource2/schemas/animlib/CNmChainSolverTask.h,  M DumpSource2/schemas/animlib/CNmClip.h,  M DumpSource2/s

+++ DumpSource2/schemas/client/CDOTA_Ability_VoidSpirit_ResonantPulse.h
@@ -2,6 +2,6 @@ class CDOTA_Ability_VoidSpirit_ResonantPulse : public C_DOTABaseAbility
 {
 	bool m_bGrantedScepterCharges;
 	float32 buff_duration;
-	int32 base_absorb_amount;
+	float32 base_absorb_amount;
 	int32 absorb_per_hero_hit;
 };
