+++ DumpSource2/schemas/client/CDOTA_Modifier_Dawnbreaker_Fire_Wreath_Attack_Bonus.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_Dawnbreaker_Fire_Wreath_Attack_Bonus : public CDOTA_Buff
 {
-	int32 attack_damage;
+	float32 attack_damage;
 	int32 creep_damage_penalty;
 };
