+++ DumpSource2/schemas/client/CDOTA_Modifier_VoidSpirit_ResonantPulse_PhysicalBuff.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_VoidSpirit_ResonantPulse_PhysicalBuff : public CDOTA_Buff
 {
-	int32 base_absorb_amount;
+	float32 base_absorb_amount;
 	int32 m_nAbsorbRemaining;
 	int32 m_nMaxAbsorb;
 	bool is_all_barrier;
