+++ DumpSource2/schemas/server/CDOTA_Modifier_BlackDragon_SplashAttack.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_BlackDragon_SplashAttack : public CDOTA_Buff
 {
-	int32 range;
+	float32 range;
 	float32 damage_percent;
 };
