+++ DumpSource2/schemas/server/CDOTA_Modifier_Furion_Sprout_Heal.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Furion_Sprout_Heal : public CDOTA_Buff
 {
 	float32 sprout_heal_interval;
-	int32 sprout_heal_per_second;
-	int32 building_healing_pct;
+	float32 sprout_heal_per_second;
+	float32 building_healing_pct;
 };
