+++ DumpSource2/schemas/server/CDOTA_Modifier_MonkeyKing_QuadrupleTap.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_MonkeyKing_QuadrupleTap : public CDOTA_Buff
 {
-	int32 counter_duration;
+	float32 counter_duration;
 	int32 required_hits;
 	int32 max_duration;
 };
