+++ DumpSource2/schemas/server/CDOTA_Modifier_Ogre_Magi_Ignite_Multicast.h
@@ -3,5 +3,5 @@ class CDOTA_Modifier_Ogre_Magi_Ignite_Multicast : public CDOTA_Buff
 	CHandle< CDOTA_BaseNPC > m_hTarget;
 	float32 multicast_delay;
 	int32 m_nMultiCastCount;
-	int32 ignite_multicast_aoe;
+	float32 ignite_multicast_aoe;
 };
