+++ DumpSource2/schemas/server/CDOTA_Modifier_TemplarAssassin_MeldArmor.h
@@ -1,4 +1,4 @@
 class CDOTA_Modifier_TemplarAssassin_MeldArmor : public CDOTA_Buff
 {
-	int32 bonus_armor;
+	float32 bonus_armor;
 };
